RefineCatDiff/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pyhd8ed1ab_0
aom=3.6.1=h59595ed_0
blas=1.0=mkl
brotli-python=1.1.0=py310hf71b8c6_2
bzip2=1.0.8=h4bc722e_7
c-ares=1.33.1=heb4867d_0
ca-certificates=2024.8.30=hbcca054_0
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py310h8deb56e_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contextlib2=21.6.0=pypi_0
contourpy=1.3.1=pypi_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.0=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.68=0
cuda-runtime=12.4.0=0
cuda-version=12.6=3
cycler=0.12.1=pypi_0
einops=0.8.0=pypi_0
expat=2.6.3=h5888daf_0
ffmpeg=4.4.2=gpl_hdf48244_113
filelock=3.16.1=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_2
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.55.2=pypi_0
freetype=2.12.1=h267a509_2
fsspec=2024.10.0=pypi_0
gettext=0.22.5=he02047a_3
gettext-tools=0.22.5=he02047a_3
gmp=6.3.0=hac33072_2
gmpy2=2.1.5=py310he8512ff_2
gnutls=3.7.9=hb077bed_0
grpcio=1.65.5=py310h1a6248f_0
h2=4.1.0=pyhd8ed1ab_0
h5py=3.12.1=pypi_0
hpack=4.0.0=pyh9f0ad1d_0
huggingface-hub=0.26.3=pypi_0
hyperframe=6.0.1=pyhd8ed1ab_0
icu=75.1=he02047a_0
idna=3.10=pyhd8ed1ab_0
importlib-metadata=8.5.0=pyha770c72_0
importlib_resources=6.4.5=pyhd8ed1ab_1
intel-openmp=2022.0.1=h06a4308_3633
jinja2=3.1.4=pyhd8ed1ab_0
kiwisolver=1.4.7=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.43=h712a8e2_1
lerc=4.0.0=h27087fc_0
libabseil=20240722.0=cxx17_h5888daf_1
libasprintf=0.22.5=he8f35ee_3
libasprintf-devel=0.22.5=he8f35ee_3
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libcublas=12.4.2.65=0
libcufft=11.2.0.44=0
libcufile=1.11.1.6=0
libcurand=10.3.7.68=0
libcusolver=11.6.0.99=0
libcusparse=12.3.0.142=0
libdeflate=1.21=h4bc722e_0
libdrm=2.4.123=hb9d3cd8_0
libegl=1.7.0=ha4b6fd6_0
libexpat=2.6.3=h5888daf_0
libffi=3.4.2=h7f98852_5
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgettextpo=0.22.5=he02047a_3
libgettextpo-devel=0.22.5=he02047a_3
libgfortran=14.1.0=h69a702a_1
libgfortran-ng=14.1.0=h69a702a_1
libgfortran5=14.1.0=hc5f4f2c_1
libgl=1.7.0=ha4b6fd6_0
libglvnd=1.7.0=ha4b6fd6_0
libglx=1.7.0=ha4b6fd6_0
libgomp=14.1.0=h77fa898_1
libgrpc=1.65.5=hf5c653b_0
libiconv=1.17=hd590300_2
libidn2=2.3.7=hd590300_0
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=16_linux64_mkl
libnpp=12.2.5.2=0
libnsl=2.0.1=hd590300_0
libnvfatbin=12.6.68=0
libnvjitlink=12.4.99=0
libnvjpeg=12.3.1.89=0
libopenblas=0.3.27=pthreads_hac2b453_1
libpciaccess=0.18=hd590300_0
libpng=1.6.44=hadc24fc_0
libprotobuf=5.27.5=h5b01275_2
libre2-11=2023.09.01=hbbce691_3
libsqlite=3.46.1=hadc24fc_0
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=14.1.0=h4852527_1
libtasn1=4.19.0=h166bdaf_0
libtiff=4.7.0=h6565414_0
libunistring=0.9.10=h7f98852_0
libuuid=2.38.1=h0b41bf4_0
libva=2.22.0=h8a09558_1
libvpx=1.13.1=h59595ed_0
libwebp-base=1.4.0=hd590300_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.12.7=he7c6b58_4
libzlib=1.3.1=h4ab18f5_1
llvm-openmp=15.0.7=h0cdce71_0
markdown=3.6=pyhd8ed1ab_0
markupsafe=2.1.5=py310ha75aee5_1
matplotlib=3.9.3=pypi_0
medpy=0.5.2=pypi_0
mkl=2022.1.0=hc2b9512_224
ml-collections=0.1.1=pypi_0
monai=1.3.2=pyhd8ed1ab_0
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_0
ncurses=6.5=he02047a_1
nettle=3.9.1=h7ab15ed_0
networkx=3.3=pyhd8ed1ab_1
nibabel=5.3.2=pyha770c72_0
numpy=2.1.1=py310hd6e36ab_0
openh264=2.3.1=hcb278e6_2
openjpeg=2.5.2=h488ebb8_0
openssl=3.4.0=hb9d3cd8_0
p11-kit=0.24.1=hc5aa10d_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.2.3=pypi_0
pillow=10.4.0=py310he228d35_1
pip=24.2=pyh8b19718_1
protobuf=5.27.5=py310hf71b8c6_0
pthread-stubs=0.4=hb9d3cd8_1002
pycparser=2.22=pyhd8ed1ab_0
pyparsing=3.2.0=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.15=h4a871b0_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.10=5_cp310
pytorch=2.4.1=py3.10_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_6
pytorch-mutex=1.0=cuda
pytz=2024.2=pypi_0
pyyaml=6.0.2=py310ha75aee5_1
re2=2023.09.01=h77b4e00_3
readline=8.2=h8228510_1
requests=2.32.3=pyhd8ed1ab_0
safetensors=0.4.5=pypi_0
scipy=1.14.1=py310ha3fb0e1_0
setuptools=75.1.0=pyhd8ed1ab_0
simpleitk=2.4.0=pypi_0
six=1.16.0=pyh6c4a22f_0
svt-av1=1.4.1=hcb278e6_0
sympy=1.13.2=pypyh2585a3b_103
tensorboard=2.18.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py310h75e40e8_1
tensorboardx=2.6.2.2=pypi_0
timm=1.0.12=pypi_0
tk=8.6.13=noxft_h4845f30_101
torch-tb-profiler=0.4.3=pypi_0
torchaudio=2.4.1=py310_cu124
torchtriton=3.0.0=py310
torchvision=0.19.1=py310_cu124
tqdm=4.66.5=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024.2=pypi_0
urllib3=2.2.3=pyhd8ed1ab_0
wayland=1.23.1=h3e06ad9_0
wayland-protocols=1.37=hd8ed1ab_0
werkzeug=3.0.4=pyhd8ed1ab_0
wheel=0.44.0=pyhd8ed1ab_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xorg-fixesproto=5.0=hb9d3cd8_1003
xorg-libx11=1.8.10=h4f16b4b_0
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxdmcp=1.1.3=hb9d3cd8_1
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-xextproto=7.3.0=hb9d3cd8_1004
xorg-xorgproto=2024.1=hb9d3cd8_1
xz=5.2.6=h166bdaf_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h7f98852_2
zipp=3.20.2=pyhd8ed1ab_0
zstandard=0.23.0=py310ha39cb0e_1
zstd=1.5.6=ha6fb4c9_0
